requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
kafka-python
scikit-learn
streamlit